Add intake drum motor support
The drum can now be toggled with the 'A' button (which adjusts
cmd_vel.angular.y because I don't want to make a custom message) but
with a fixed speed.  Works well!

I also increased maximum throttle for the drive motors to 50%, which
is a bit... excessive.  We probably want to reduce back to 25%.